refactor: organize modules into per-feature subfolders, standardize imports
