+++ packages/core/src/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { VERSION } from './index.js';
+import { VERSION } from './index';
 
 describe('@humanjs/core', () => {
   it('exports VERSION', () => {
