+++ packages/core/src/presets/index.ts
@@ -1,4 +1,4 @@
-export { careful } from './careful.js';
-export { distracted } from './distracted.js';
-export { fast } from './fast.js';
-export { precise } from './precise.js';
+export { careful } from './careful';
+export { distracted } from './distracted';
+export { fast } from './fast';
+export { precise } from './precise';
