+++ packages/playwright/src/mouse/click.test.ts
@@ -1,6 +1,6 @@
 import type { Locator, Page } from 'playwright';
 import { describe, expect, it, vi } from 'vitest';
-import { createHuman } from './index.js';
+import { createHuman } from '../index';
 
 interface MockBoundingBox {
   readonly x: number;
